feat: pair session ID with device data and configure environments and time interval
