+++ Airwallex.podspec
@@ -22,7 +22,7 @@ Pod::Spec.new do |s|
     plugin.source_files = 'Airwallex/Core/Sources/**/*.{swift,h,m}'
     plugin.public_header_files = 'Airwallex/Core/Sources/*.h'
     plugin.resources = ['Airwallex/Core/Resources/AirwallexCore.bundle']
-    plugin.vendored_frameworks = 'Frameworks/AirTracker.xcframework', 'Frameworks/AirRisk.xcframework'
+    plugin.vendored_frameworks = 'Frameworks/AirTracker.xcframework', 'Frameworks/AirwallexRisk.xcframework'
   end
   
   s.subspec 'WeChatPay' do |plugin|
