--- Frameworks/AirRisk.xcframework/ios-arm64/AirRisk.framework/Headers/AirRisk.h
@@ -1,18 +1,0 @@
-//
-//  AirRisk.h
-//  AirRisk
-//
-//  Created by Hector.Huang on 2024/7/9.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for AirRisk.
-FOUNDATION_EXPORT double AirRiskVersionNumber;
-
-//! Project version string for AirRisk.
-FOUNDATION_EXPORT const unsigned char AirRiskVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <AirRisk/PublicHeader.h>
-
-
