--- Frameworks/AirRisk.xcframework/ios-arm64/AirRisk.framework/Modules/AirRisk.swiftmodule/arm64-apple-ios.private.swiftinterface
@@ -1,47 +1,0 @@
-// swift-interface-format-version: 1.0
-// swift-compiler-version: Apple Swift version 5.10 (swiftlang-5.10.0.13 clang-1500.3.9.4)
-// swift-module-flags: -target arm64-apple-ios13.0 -enable-objc-interop -enable-library-evolution -swift-version 5 -enforce-exclusivity=checked -O -enable-bare-slash-regex -module-name AirRisk
-@_exported import AirRisk
-import Foundation
-import Swift
-import UIKit
-import _Concurrency
-import _StringProcessing
-import _SwiftConcurrencyShims
-@_hasMissingDesignatedInitializers @objc public class AirwallexRisk : ObjectiveC.NSObject {
-  @objc deinit
-}
-extension AirRisk.AirwallexRisk {
-  @objc public static func start(accountID: Swift.String?, with configuration: AirRisk.AirwallexRiskConfiguration = .init())
-  @objc public static func set(accountID: Swift.String?)
-  @objc public static func set(userID: Swift.String?)
-  @objc public static func log(event: Swift.String, screen: Swift.String? = nil)
-  public static var header: AirRisk.Header? {
-    get
-  }
-}
-public struct Header {
-  public let field: Swift.String
-  public let value: Swift.String
-}
-extension Foundation.URLRequest {
-  public mutating func setAirwallexHeader()
-}
-@objc public class AirwallexRiskConfiguration : ObjectiveC.NSObject {
-  @objc public init(isProduction: Swift.Bool = true, tenant: AirRisk.Tenant = .scale)
-  @objc deinit
-}
-@objc public enum Tenant : Swift.Int, Swift.Codable {
-  case `internal`
-  case pa
-  case scale
-  public func encode(to encoder: any Swift.Encoder) throws
-  public init?(rawValue: Swift.Int)
-  public typealias RawValue = Swift.Int
-  public var rawValue: Swift.Int {
-    get
-  }
-}
-extension AirRisk.Tenant : Swift.Equatable {}
-extension AirRisk.Tenant : Swift.Hashable {}
-extension AirRisk.Tenant : Swift.RawRepresentable {}
