--- Frameworks/AirRisk.xcframework/ios-arm64/AirRisk.framework/Modules/module.modulemap
@@ -1,11 +1,0 @@
-framework module AirRisk {
-  umbrella header "AirRisk.h"
-  export *
-
-  module * { export * }
-}
-
-module AirRisk.Swift {
-  header "AirRisk-Swift.h"
-  requires objc
-}
