--- Frameworks/AirRisk.xcframework/ios-arm64_x86_64-simulator/AirRisk.framework/_CodeSignature/CodeResources
@@ -1,312 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>files</key>
-	<dict>
-		<key>Headers/AirRisk-Swift.h</key>
-		<data>
-		XVFJO4nnKv31rKcScremq7FPiVU=
-		</data>
-		<key>Headers/AirRisk.h</key>
-		<data>
-		3no9x433kseGeQJHZ9jnMKkBHLc=
-		</data>
-		<key>Info.plist</key>
-		<data>
-		L6mNCDKLXqMM6l02B77Ta0OJWFY=
-		</data>
-		<key>Modules/AirRisk.swiftmodule/arm64-apple-ios-simulator.abi.json</key>
-		<data>
-		koOTngti2jXHORkNgS5kBXExiMI=
-		</data>
-		<key>Modules/AirRisk.swiftmodule/arm64-apple-ios-simulator.private.swiftinterface</key>
-		<data>
-		TKosm5m+uCWsJtT3mZs/VLiA/LY=
-		</data>
-		<key>Modules/AirRisk.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
-		<data>
-		roeKCOKAH7jQsMM7sFdlBVimuAE=
-		</data>
-		<key>Modules/AirRisk.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
-		<data>
-		TKosm5m+uCWsJtT3mZs/VLiA/LY=
-		</data>
-		<key>Modules/AirRisk.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
-		<data>
-		XyNJDhvS651utDdRcxdIzE/m0zk=
-		</data>
-		<key>Modules/AirRisk.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
-		<data>
-		koOTngti2jXHORkNgS5kBXExiMI=
-		</data>
-		<key>Modules/AirRisk.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
-		<data>
-		2j6Wkcti2JcP9k09gOANW8R3G0c=
-		</data>
-		<key>Modules/AirRisk.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
-		<data>
-		Bbto9H821QY8NbJQSngO47BL9SU=
-		</data>
-		<key>Modules/AirRisk.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
-		<data>
-		2j6Wkcti2JcP9k09gOANW8R3G0c=
-		</data>
-		<key>Modules/AirRisk.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
-		<data>
-		8ZpX5fycMHTQvc9aQ3QA9iCGNJY=
-		</data>
-		<key>Modules/module.modulemap</key>
-		<data>
-		u/jBuwOnszhL/I6DYXLUCkL4dPM=
-		</data>
-		<key>version.json</key>
-		<data>
-		5jXOZYydr63x/6RfC1vjtwJ+nFM=
-		</data>
-	</dict>
-	<key>files2</key>
-	<dict>
-		<key>Headers/AirRisk-Swift.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			XVFJO4nnKv31rKcScremq7FPiVU=
-			</data>
-			<key>hash2</key>
-			<data>
-			pErgvXX97MDb3IkcnwSgOSQ4Cmv0QKofRwbR2r62y8I=
-			</data>
-		</dict>
-		<key>Headers/AirRisk.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			3no9x433kseGeQJHZ9jnMKkBHLc=
-			</data>
-			<key>hash2</key>
-			<data>
-			F0tZuZGRyiHJUvJHlGrIRWe+W6dwsXykIjVCrcYZY6g=
-			</data>
-		</dict>
-		<key>Modules/AirRisk.swiftmodule/arm64-apple-ios-simulator.abi.json</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			koOTngti2jXHORkNgS5kBXExiMI=
-			</data>
-			<key>hash2</key>
-			<data>
-			hfFRTCayy9uZdLiWDq9rnfwgtkjATCTmcape6TeC0fs=
-			</data>
-		</dict>
-		<key>Modules/AirRisk.swiftmodule/arm64-apple-ios-simulator.private.swiftinterface</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			TKosm5m+uCWsJtT3mZs/VLiA/LY=
-			</data>
-			<key>hash2</key>
-			<data>
-			1L3y2OkZJli1ACSM3KL7tOMrA+PBXVolv5Bn1Qd47wk=
-			</data>
-		</dict>
-		<key>Modules/AirRisk.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			roeKCOKAH7jQsMM7sFdlBVimuAE=
-			</data>
-			<key>hash2</key>
-			<data>
-			X+lL/YmCFjRQeoFl6AWPvCK2n8aIxwYoATzWaZ7Z4xc=
-			</data>
-		</dict>
-		<key>Modules/AirRisk.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			TKosm5m+uCWsJtT3mZs/VLiA/LY=
-			</data>
-			<key>hash2</key>
-			<data>
-			1L3y2OkZJli1ACSM3KL7tOMrA+PBXVolv5Bn1Qd47wk=
-			</data>
-		</dict>
-		<key>Modules/AirRisk.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			XyNJDhvS651utDdRcxdIzE/m0zk=
-			</data>
-			<key>hash2</key>
-			<data>
-			U0MzZX7ktPBKE1B52quotwCDCpyou/sALHfgu7SQPw8=
-			</data>
-		</dict>
-		<key>Modules/AirRisk.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			koOTngti2jXHORkNgS5kBXExiMI=
-			</data>
-			<key>hash2</key>
-			<data>
-			hfFRTCayy9uZdLiWDq9rnfwgtkjATCTmcape6TeC0fs=
-			</data>
-		</dict>
-		<key>Modules/AirRisk.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			2j6Wkcti2JcP9k09gOANW8R3G0c=
-			</data>
-			<key>hash2</key>
-			<data>
-			vU3p/bqoV2DkwLrJur45FCyIdMMeGf4QHGofVSqeD5E=
-			</data>
-		</dict>
-		<key>Modules/AirRisk.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			Bbto9H821QY8NbJQSngO47BL9SU=
-			</data>
-			<key>hash2</key>
-			<data>
-			ukUcCVmfC7vZeVFUxQ76lZZE2U5mJrZmtxP/6FPwlCQ=
-			</data>
-		</dict>
-		<key>Modules/AirRisk.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			2j6Wkcti2JcP9k09gOANW8R3G0c=
-			</data>
-			<key>hash2</key>
-			<data>
-			vU3p/bqoV2DkwLrJur45FCyIdMMeGf4QHGofVSqeD5E=
-			</data>
-		</dict>
-		<key>Modules/AirRisk.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			8ZpX5fycMHTQvc9aQ3QA9iCGNJY=
-			</data>
-			<key>hash2</key>
-			<data>
-			dryGPdtE+p3AoK360wpOtiL7RqEkw4TWsjex7zm3FmA=
-			</data>
-		</dict>
-		<key>Modules/module.modulemap</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			u/jBuwOnszhL/I6DYXLUCkL4dPM=
-			</data>
-			<key>hash2</key>
-			<data>
-			gCgJHwClyFz3O0Oz4WdyeFE/u5V3i/+JcUF4z2s6vNQ=
-			</data>
-		</dict>
-		<key>version.json</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			5jXOZYydr63x/6RfC1vjtwJ+nFM=
-			</data>
-			<key>hash2</key>
-			<data>
-			MupMZPCX6aTI8IZtupGTAHTZBXL8+SwH2KjycSnQx1I=
-			</data>
-		</dict>
-	</dict>
-	<key>rules</key>
-	<dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^version.plist$</key>
-		<true/>
-	</dict>
-	<key>rules2</key>
-	<dict>
-		<key>.*\.dSYM($|/)</key>
-		<dict>
-			<key>weight</key>
-			<real>11</real>
-		</dict>
-		<key>^(.*/)?\.DS_Store$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>2000</real>
-		</dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^Info\.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^PkgInfo$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^embedded\.provisionprofile$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^version\.plist$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-	</dict>
-</dict>
-</plist>
