+++ Frameworks/AirwallexRisk.xcframework/ios-arm64_x86_64-simulator/AirwallexRisk.framework/_CodeSignature/CodeResources
@@ -1,0 +1,223 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/AirwallexRisk-Swift.h</key>
+		<data>
+		fyttjc7JzM63G7i0ty9Zwy+iNrs=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		C2sstsfUn5jI2tXbZOKPeUL52+I=
+		</data>
+		<key>Modules/AirwallexRisk.swiftmodule/arm64-apple-ios-simulator.abi.json</key>
+		<data>
+		r7W1RvKMwbZwo2nD9wEVcCtp8C4=
+		</data>
+		<key>Modules/AirwallexRisk.swiftmodule/arm64-apple-ios-simulator.private.swiftinterface</key>
+		<data>
+		2KfvPBLvssw4Gr8Hul6opbpQnIk=
+		</data>
+		<key>Modules/AirwallexRisk.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
+		<data>
+		HoHoBTt6pFW6RLQkYFGJT83pLKU=
+		</data>
+		<key>Modules/AirwallexRisk.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
+		<data>
+		2KfvPBLvssw4Gr8Hul6opbpQnIk=
+		</data>
+		<key>Modules/AirwallexRisk.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
+		<data>
+		fKzrOP0wcqTsuKlR3jr6TmRxT9I=
+		</data>
+		<key>Modules/AirwallexRisk.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
+		<data>
+		r7W1RvKMwbZwo2nD9wEVcCtp8C4=
+		</data>
+		<key>Modules/AirwallexRisk.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
+		<data>
+		hKpGEl7k49wanEwOCqLjb0sMYlY=
+		</data>
+		<key>Modules/AirwallexRisk.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<data>
+		++TjSrl9X7PHiiwOwOxRPWNHZXY=
+		</data>
+		<key>Modules/AirwallexRisk.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<data>
+		hKpGEl7k49wanEwOCqLjb0sMYlY=
+		</data>
+		<key>Modules/AirwallexRisk.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<data>
+		pCAvFFn3PmMUSJTXKe/euNnlMbY=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/AirwallexRisk-Swift.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			mt6nIGSfZRAtFOIkvk6WgTS6lV3DEzoFIfGpvH+EmI4=
+			</data>
+		</dict>
+		<key>Modules/AirwallexRisk.swiftmodule/arm64-apple-ios-simulator.abi.json</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			JrMJWkUUD5ppHJ8Sk2u88jmWgezju4X2pcvBlyqeKmo=
+			</data>
+		</dict>
+		<key>Modules/AirwallexRisk.swiftmodule/arm64-apple-ios-simulator.private.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			b/3PUB5zKZesXeYsABX0qiPipPLdotidTtNsPSqV+yQ=
+			</data>
+		</dict>
+		<key>Modules/AirwallexRisk.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			iZZmEJDlkkyT1OXEHtq2L4Sc3XKkqH1ZPKDIss4l7+c=
+			</data>
+		</dict>
+		<key>Modules/AirwallexRisk.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			b/3PUB5zKZesXeYsABX0qiPipPLdotidTtNsPSqV+yQ=
+			</data>
+		</dict>
+		<key>Modules/AirwallexRisk.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			KYaXCPRd99GfaOlhQb1weckQ+Jre1yHMPGl7LQb3pdA=
+			</data>
+		</dict>
+		<key>Modules/AirwallexRisk.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			JrMJWkUUD5ppHJ8Sk2u88jmWgezju4X2pcvBlyqeKmo=
+			</data>
+		</dict>
+		<key>Modules/AirwallexRisk.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			ZHbuX/5c4NkdiJtQvylEnxgqyWMjgEU4jLWLzEnAjek=
+			</data>
+		</dict>
+		<key>Modules/AirwallexRisk.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			pb8MW1NJZwdhg1EXG1KuuVIuN+ANUIN4+WV9bt6XP+U=
+			</data>
+		</dict>
+		<key>Modules/AirwallexRisk.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			ZHbuX/5c4NkdiJtQvylEnxgqyWMjgEU4jLWLzEnAjek=
+			</data>
+		</dict>
+		<key>Modules/AirwallexRisk.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			1kQqtcsHRfZP1lSIgsJW6fv50qbAsgdDcbh1/6xYaJ8=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
